NavTransitionDirectionContext to a separate folder
