+++ packages/vkui/src/components/NavTransitionContext/NavTransitionContext.tsx
@@ -22,33 +22,3 @@ export const NavTransitionProvider = ({
   });
   return <TransitionContext.Provider value={contextValue}>{children}</TransitionContext.Provider>;
 };
-
-const TransitionDirectionContext = React.createContext<boolean | undefined>(undefined);
-
-export const NavTransitionDirectionProvider = ({
-  children,
-  isBack: isBackProp,
-}: React.PropsWithChildren<{ isBack: boolean | undefined }>) => {
-  const parentIsBack = React.useContext(TransitionDirectionContext);
-  // if local isBack is undefined then transition happend on the parent side (probably Root)
-  const isBack = isBackProp !== undefined ? isBackProp : parentIsBack;
-
-  // 'direction' should always represent the direction state of the panel on mount
-  // save the on mount value of the panel to the state
-  // to make sure we don't trigger new re-render for the panel
-  // due to change in the prop passed to provider
-  const [isBackOnMount] = React.useState<boolean | undefined>(isBack);
-
-  return (
-    <TransitionDirectionContext.Provider value={isBackOnMount}>
-      {children}
-    </TransitionDirectionContext.Provider>
-  );
-};
-
-export const useNavDirection = () => {
-  const isBack = React.useContext(TransitionDirectionContext);
-  const transitionDirection = isBack === undefined ? undefined : isBack ? 'backwards' : 'forwards';
-
-  return transitionDirection;
-};
